Refactor code to include auth.py. Move all auth views to this view.
